pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "backend проекта Особенный тип"
authors = ["Константин Трохимец <katrohimets@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.21.0"
django = "3.2.16"
pydantic = "^1.10.2"
yookassa = "^2.3.2"
djangorestframework = "^3.14.0"
requests = "^2.28.1"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.4"
pillow = "^9.3.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
django-cors-headers = "^3.13.0"
django-rest-captcha = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
